+++ js/consts.js
@@ -1,0 +1,30 @@
+const NAMES = [
+  'Магомед',
+  'Александра',
+  'Евгения',
+  'Артем',
+  'Варвара',
+  'Илья',
+];
+
+const DESCRIPTION = [
+  'Красивое описание фотографии',
+  'Минимум вложений и максимум пользы. Если что, на фото моя собака',
+  'Мое утро',
+  'Как говорится, без кота и жизнь не та',
+];
+
+const COMMENTS_COUNT_MAX = 6;
+const COMMENTS_COUNT_MIN = 2;
+const result = [];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {NAMES, DESCRIPTION, COMMENTS_COUNT_MAX, COMMENTS_COUNT_MIN, result, COMMENTS};
